requirements.txt
appdirs==1.4.0
Django==1.10.4
django-allauth==0.32.0
django-bootstrap3==8.2.3
django-datatable==0.2.1
django-filter==1.0.1
django-form-utils==1.0.3
django-jet==1.0.4
google-api-python-client==1.6.1
gunicorn==19.6.0
httplib2==0.9.2
Markdown==2.6.7
oauth2client==4.0.0
oauthlib==2.0.2
packaging==16.8
pyasn1==0.1.9
pyasn1-modules==0.0.8
pyparsing==2.1.10
python-http-client==2.2.1
python-openid==2.2.5
requests==2.12.3
requests-oauthlib==0.8.0
rsa==3.4.2
sendgrid==3.6.3
sendgrid-django==4.0.4
six==1.10.0
uritemplate==3.0.0
uuid==1.30
